[Refactor] Extract APICondition into a new file
