+++ iOS ExerciseTests/Content/Helpers/DataLoaderMock.swift
@@ -8,14 +8,6 @@
 import UIKit
 @testable import iOS_Exercise
 
-enum APICondition {
-    case successWithJSON
-    case successWithJSONButNoData
-    case successWithImage
-    case networkFailure
-    case decodeFailure
-}
-
 class DataLoaderMock: DataLoaderProtocol {
     
     private var currentRequestIndex: Int = 0
